requirements.txt
Flask
requests
slackclient
slackeventsapi
urllib3
prompt-toolkit
